pyproject.toml
[project]
name = "bigbrain"
version = "0.1.0"
description = "A knowledge compiler that ingests, distills, and compiles knowledge from multiple sources."
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "pymupdf>=1.23",
    "pypdf>=3.0",
    "httpx>=0.27",
    "notion-client>=2.0",
    "beautifulsoup4>=4.12",
    "html2text>=2024.2",
    "tenacity>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
postgres = ["psycopg[binary]>=3.1", "pgvector>=0.2"]
neo4j = ["neo4j>=5.0"]
qdrant = ["qdrant-client>=1.7"]
weaviate = ["weaviate-client>=4.0"]
pinecone = ["pinecone>=3.0"]
all-stores = [
    "psycopg[binary]>=3.1", "pgvector>=0.2",
    "neo4j>=5.0",
    "qdrant-client>=1.7",
    "weaviate-client>=4.0",
    "pinecone>=3.0",
]

[project.scripts]
bigbrain = "bigbrain.cli:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
